mcp_a2a_master/pyproject.toml
[project]
name = "mcp-a2a-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.15",
    "asyncclick>=8.1.8",
    "google-adk>=1.7.0",
    "mcp[cli]>=1.12.0",
    "openai-agents>=0.0.7",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]
